app/src/test/java/com/app/chinwag/repository/ChangePasswordRepositoryTest.kt
package com.app.chinwag.repository

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import com.app.chinwag.dataclasses.generics.TAListResponse
import com.app.chinwag.objectclasses.KotlinBaseMockObjectsClass
import com.app.chinwag.utils.mock
import com.app.chinwag.utils.whenever
import io.reactivex.Single
import junit.framework.TestCase.assertEquals
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.mockito.ArgumentMatchers
import org.mockito.Mockito

class ChangePasswordRepositoryTest : KotlinBaseMockObjectsClass() {

    @Rule
    @JvmField
    val run = InstantTaskExecutorRule()

    private val mockChangePasswordRepositoryTest = mock<ChangePasswordRepository>()
    private val changePasswordRepository by lazy {
        ChangePasswordRepository(mockNetworkService)
    }

    @Before
    fun setUp() {
        Mockito.reset(mockNetworkService, mockApplication)
    }

    @Test
    fun verifyConstructorParameter(){
        assertEquals(mockNetworkService, changePasswordRepository.networkService)
    }

    @Test
    fun callChangePassword(){
        whenever(mockNetworkService.callChangePassword("Pass@123","9545954400"))
            .thenReturn(Single.just(TAListResponse()))
        whenever(mockChangePasswordRepositoryTest.callChangePassword(ArgumentMatchers.anyString(), ArgumentMatchers.anyString()))
            .thenReturn(Single.just(TAListResponse()))
        changePasswordRepository.callChangePassword("Pass@123","9545954400")
            .test().assertComplete()
    }
}